requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.0.0
alembic
python-dotenv
sqlalchemy
psycopg2-binary
pydantic-settings
email-validator
passlib
python-jose[cryptography]
python-multipart
redis
slowapi